requirements.txt
pip~=23.2.1
attrs~=24.2.0
typing_extensions~=4.12.2
idna~=3.10
multidict~=6.1.0
propcache~=0.2.0
aiogram~=3.13.1
pydantic~=2.9.2
aiofiles~=24.1.0
magic-filter~=1.0.12
certifi~=2024.8.30
aiohttp~=3.10.10
yarl~=1.15.0
aiosignal~=1.3.1
frozenlist~=1.4.1
aiohappyeyeballs~=2.4.3
pydantic_core~=2.23.4
annotated-types~=0.7.0
python-dotenv~=1.0.1
requests~=2.32.3